Fix trotter step count type: numpy.float64 cannot repeat a list
_n_trotter_steps returned np.ceil(...) directly, a numpy.float64, and
get_trotter_program builds its subroutine sequence with [0] * steps,
which raises TypeError for non-int. The trotter path appears never to
have been exercised upstream (ex_2 demonstrates QSP only; the trotter
import is unused there). Found wiring benchq-mcp estimate_time_evolution
to trotter_time_evolution_algorithm; with this cast the trotter path
estimates end to end on native Windows.

+++ src/benchq/algorithms/time_evolution.py
@@ -56,7 +56,9 @@ def qsp_time_evolution_algorithm(
 # TODO: This method of calculating number of steps is not exact.
 # It doesn't take into account the prefactor coming from the Hamiltonian.
 def _n_trotter_steps(evolution_time, total_trotter_error) -> int:
-    return np.ceil(evolution_time / total_trotter_error)
+    # int() matters: get_trotter_program repeats its subroutine sequence
+    # with [0] * steps, and numpy.float64 cannot repeat a list.
+    return int(np.ceil(evolution_time / total_trotter_error))
 
 
 def trotter_time_evolution_algorithm(
